Move hardware_type into device.rs
