+++ src/models/mod.rs
@@ -1,6 +1,5 @@
 pub mod device;
 pub mod device_type;
-pub mod hardware_type;
 pub mod tv;
 
 mod tests;
